Add a version without horizontal data component rotation
git-svn-id: file:///home/svn/svn/wphase_project/trunk@13 cdf9f54e-c1c6-40d2-88f3-cdd711570eaa

+++ src/FAST_SYNTH/rotate_traces.c
@@ -2,13 +2,13 @@
 #include <stdlib.h>
 
 void 
-rotate_traces(float *T, float *P, float baz, int npts, float *N, float *E)
+rotate_traces(double *T, double *P, float baz, int npts, double *N, double *E)
 {
   double co, si;
   int j;
 
-  co = cos(M_PI*baz/180.);
-  si = sin(M_PI*baz/180.);
+  co = cos(M_PI*(double)baz/180.);
+  si = sin(M_PI*(double)baz/180.);
 
   for(j=0; j<npts; j++)
 	{
